requirements.txt
Flask
qrcode[pil]
pillow